Merged PR 14287: Versioning and Plugin Administration
Each iModel.js system module now reveals its SEMVER build number as it loads.
When a Plugin is built, it records the SEMVER build numbers of each module in the dependencies of its package, and it checks to verify that compatible versions are available at runtime.

Webpack will now insert the required iModel.js system modules into an index.html template. The IModelJsLoader now looks at the required modules and decides what to load, rather than looking at the data-uicomponents and data-uiframework arguments passed to it (those are now obsolete).

The presentation-* modules are now treated as a separate group from ui-framework.