Merged PR 37737: Presentation: property grid records with links and default behavior on click
Added onPropertyLinkClick property to PropertyGrid component with default behavior to open url links in a new tab or launch default email client for email links if it is not set. Content builder now checks if records have links in them and if so sets links property for them.

Related work items: #145868