Presentation: Make tree component auto update more robust (#2001)
The old way to perform tree hierarchy update on client side relied on comparing the result of two presentation queries on the backend. It involved loading all nodes that are potentially visible to the user, which is not ideal when working with enormous tree structures. The implementation was also very complicated and contained many hidden bugs.

The new way of performing tree update after hierarchy change does all the work in the frontend. Not only this is much easier to implement, we make it way faster by only loading a fraction of possibly visible nodes. We achieve this by recording which nodes are expanded at the time of update and then attempting to reconstruct the expanded tree structure in the updated tree. The tradeoff here is that we might not be able to find previously expanded nodes in the updated hierarchy because they might have moved up or down among their siblings. We make a few attempts to guess the new locations of these nodes, and most of the time that is enough to find the vast majority of them and create an impression to the user that the expanded tree hierarchy is kept after the update.